forge fmt

+++ test/handlers/DelegateTokenHandler.t.sol
@@ -71,7 +71,7 @@ contract DelegateTokenHandler is CommonBase, StdCheats, StdUtils {
         delegateToken = IDelegateToken(dt);
         principal = PrincipalToken(IDelegateToken(dt).principalToken());
         for (uint256 i; i < TOTAL_TOKENS; i++) {
-            tokenContracts.add(address(new MockERC721(uint(keccak256(abi.encode("start_id", i))))));
+            tokenContracts.add(address(new MockERC721(uint256(keccak256(abi.encode("start_id", i))))));
         }
     }
